add getAllPetProviders to PetProviders Service.
