--- backend/src/main/java/com/SegundasHuellas/backend/petProviders/internal/application/service/PetProviderRegistrationService.java
@@ -1,8 +1,0 @@
-package com.SegundasHuellas.backend.petProviders.internal.application.service;
-
-import com.SegundasHuellas.backend.auth.api.dto.AuthenticationResponse;
-import com.SegundasHuellas.backend.petProviders.internal.application.dto.PetProviderRegistrationRequest;
-
-public interface PetProviderRegistrationService {
-    AuthenticationResponse registerPetProvider(PetProviderRegistrationRequest request);
-}
