+++ backend/src/main/java/com/SegundasHuellas/backend/petProviders/internal/infra/web/PetProviderAuthController.java
@@ -3,7 +3,7 @@ package com.SegundasHuellas.backend.petProviders.internal.infra.web;
 
 import com.SegundasHuellas.backend.auth.api.dto.AuthenticationResponse;
 import com.SegundasHuellas.backend.petProviders.internal.application.dto.PetProviderRegistrationRequest;
-import com.SegundasHuellas.backend.petProviders.internal.application.service.PetProviderRegistrationService;
+import com.SegundasHuellas.backend.petProviders.internal.application.service.PetProviderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class PetProviderAuthController {
 
-    private final PetProviderRegistrationService registrationService;
+    private final PetProviderService registrationService;
 
     @PostMapping("/register")
     public AuthenticationResponse registerPetProvider(PetProviderRegistrationRequest request) {
